Fix client side error checking for make ticket
Previous code did not do anything at all as the form had no ID.

Previous code once the ID was added still did not do anything as it did not wait for the page to load.

Previous code once updated to wait was busted as it did not actually check for important stuff such as if values were blank, some non required values failed as the js thought they were required, and multiple instances of undefined vars being called.